feature: create more features

+++ api/src/controllers/StudentsController.ts
@@ -3,16 +3,34 @@ import { validationResult } from "express-validator";
 
 import { apiMessage } from "../helpers/error";
 import { StudentsRepository } from "../repositories/StudentsRepository";
-import { prisma } from "../database/client";
 
 export class StudentsController {
   async index(req: Request, res: Response) {
+    const errors = validationResult(req);
     const { user_id } = req;
 
+    if (!errors.isEmpty()) {
+      return res
+        .status(422)
+        .send(apiMessage(false, 422, "Validation error", errors.array()));
+    }
+    const filters = {
+      ra: req.query.ra as string,
+      name: req.query.name as string,
+      cpf: req.query.cpf as string,
+      order: req.query.order as string,
+      sort: req.query.sort as string,
+    };
     const studentsRepository = new StudentsRepository();
-    const students = await studentsRepository.index(Number(user_id));
 
-    return res.send(apiMessage(true, 200, "Students", students));
+    try {
+      const students = await studentsRepository.listByUserId(Number(user_id), filters);
+
+      return res.send(apiMessage(true, 200, "Students", students));
+    } catch (error) {
+      return res.status(500)
+        .send(apiMessage(false, 500, "Ocorreu um erro ao buscar os alunos"));
+    }
   }
 
   async create(req: Request, res: Response) {
@@ -32,21 +50,28 @@ export class StudentsController {
     }
 
     const studentsRepository = new StudentsRepository();
-    const findStudent = await studentsRepository.findByRa(ra);
 
-    if (findStudent) {
-      return res.status(400).send(apiMessage(false, 400, "Aluno já cadastrado", null));
+    try {
+      const findStudent = await studentsRepository.findByRa(ra);
+
+      if (findStudent) {
+        return res.status(400)
+          .send(apiMessage(false, 400, "Aluno já cadastrado"));
+      }
+
+      const student = await studentsRepository.create({
+        name,
+        user_id: Number(user_id),
+        email,
+        cpf,
+        ra,
+      });
+
+      return res.send(apiMessage(true, 201, "Aluno criado", student));
+    } catch (error) {
+      return res.status(500)
+        .send(apiMessage(false, 500, "Ocorreu um erro ao criar o aluno"));
     }
-
-    const student = await studentsRepository.create({
-      name,
-      user_id: Number(user_id),
-      email,
-      cpf,
-      ra,
-    });
-
-    return res.send(apiMessage(true, 201, "Aluno criado", student));
   }
 
   async show(req: Request, res: Response) {
@@ -56,11 +81,12 @@ export class StudentsController {
     const studentsRepository = new StudentsRepository();
 
     try {
-      const student = await studentsRepository.show(Number(id), Number(user_id));
+      const student = await studentsRepository.findByIdAndUserId(Number(id), Number(user_id));
 
       return res.send(apiMessage(true, 200, "Student", student));
     } catch (error) {
-      return res.status(500).send(apiMessage(false, 500, "Ocorreu um erro ao buscar o aluno", null));
+      return res.status(500)
+        .send(apiMessage(false, 500, "Ocorreu um erro ao buscar o aluno"));
     }
   }
 
@@ -80,14 +106,27 @@ export class StudentsController {
     }
 
     const studentsRepository = new StudentsRepository();
-    const student = await studentsRepository.update({
-      name,
-      user_id: Number(user_id),
-      email,
-      id: Number(id),
-    });
 
-    return res.send(apiMessage(true, 200, "Aluno atualizado", student));
+    try {
+      const findStudent = await studentsRepository.findByIdAndUserId(Number(id), Number(user_id));
+
+      if (!findStudent) {
+        return res.status(404)
+          .send(apiMessage(false, 404, "Aluno não encontrado"));
+      }
+
+      const student = await studentsRepository.update({
+        name,
+        user_id: Number(user_id),
+        email,
+        id: Number(id),
+      });
+
+      return res.send(apiMessage(true, 200, "Aluno atualizado", student));
+    } catch (error) {
+      return res.status(500)
+        .send(apiMessage(false, 500, "Ocorreu um erro ao atualizar o aluno"));
+    }
   }
 
   async delete(req: Request, res: Response) {
@@ -95,11 +134,24 @@ export class StudentsController {
     const { user_id } = req;
 
     const studentsRepository = new StudentsRepository();
-    const student = await studentsRepository.delete(
-      Number(id),
-      Number(user_id),
-    );
 
-    return res.send(apiMessage(true, 200, "Aluno deletado", student));
+    try {
+      const findStudent = await studentsRepository.findByIdAndUserId(Number(id), Number(user_id));
+
+      if (!findStudent) {
+        return res.status(404)
+          .send(apiMessage(false, 404, "Aluno não encontrado"));
+      }
+
+      const student = await studentsRepository.delete(
+        Number(id),
+        Number(user_id),
+      );
+
+      return res.send(apiMessage(true, 200, "Aluno deletado", student));
+    } catch (error) {
+      return res.status(500)
+        .send(apiMessage(false, 500, "Ocorreu um erro ao deletar o aluno"));
+    }
   }
 }
